Time: 4 ms (20.37%), Space: 62.9 MB (48.9%) - LeetHub

+++ 0011-container-with-most-water/0011-container-with-most-water.cpp
@@ -17,6 +17,6 @@ public:
                 right--;
             }
         }
-        return ans;
+        return (int)ans;
     }
 };
